requirements.txt
tensorflow-cpu==2.16.1
streamlit==1.8.1
streamlit-drawable-canvas
numpy
scikit-image
protobuf==3.20.3
